Add files via upload
The Sorting Visualizer project is a web development initiative designed to provide a visual representation of various sorting algorithms, such as Bubble Sort, Selection Sort, and Merge Sort. The primary goal is to facilitate learning and understanding of how these algorithms operate by allowing users to observe the step-by-step execution of sorting operations.

The project features a graphical representation of data in the form of bars, initially arranged in a random, unsorted order. Users can initiate the sorting algorithms by clicking on corresponding buttons for each algorithm. As the algorithms progress, the bars undergo swapping or other operations, showcasing the underlying mechanisms of each sorting method.

The main objective of Sorting Visualizer project is to learn and visualise the sorting algorithms how these algorithms works We can see the steps of swapping or any divide and conquer part of these bars. Here 5 different types of sorting algorithms is used to visualise.  A visualization of data is implemented as a bar graph, after which a data sorting and algorithm may be applied .The resulting animation is then performed either automatically or by the user, who then sets their own pace.